backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
openai>=1.35.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
numpy>=1.26.0
python-dotenv>=1.0.0
